Fix unintended node selection bug and tooltip dismissal bug
This fixes two bugs:
* Unintended skill selection while panning -- The problem here is that
if you click on a node and then drag, the mouse up event causes the drag
to end but it also causes the node to get selected.  The fix is to mark
the up event as handled if we dragged a meaningful distance; this
prevents the up event from also selecting the node.  This bug gets hit
all the time when I watch streamers and is the common root cause for
"Hey why did you select node XYZ?" "Oh whoops, I don't know how that
happened"

* Tooltips don't get dismissed sometimes -- If you have another window
open over the planner and then mouse over a node, a tooltip will pop up.
If you do it fast enough such that you just briefly move over a node
before your mouse ends up in another window, you can end up with a
tooltip that never goes away.  The fix here is to dismiss the tooltip
whenever the mouse leaves the canvas.