More tweaking
Main change in this commit was to greatly reduce the search space in
"normal" inputs (that is, target nodes are somewhat evenly spaced out).

I want to investigate employing an simulated annealing-style mutation in
the GA. It should work well since the cost function should be pretty
favorably-shaped over the search space.