Added Crossed HighlightState
Added another HighlightState for nodes that is accessible with middle
mouse click.
Crossed nodes get ignored by the optimizer and
SkillTree.GetShortestPathTo.
If omitting nodes disconnect the tree for the optimizer, a popup is
shown (Visual Studio stops after the exception that gets handled later,
so to see the popup, press F5).