.NET 4.5, custom Settings path
Changed target framework to .NET 4.5 to be able to use async/await for
cleaner and easier background tasks (see MainWindow.xaml.cs). That also
meant redownloading all the packages.
Also implemented a CustomSettingsProvider to be able to store the
WindowPosition in a custom directory. (Which is the folder path returned
by AppData,GetFolder("Settings")). This needed an update of
MahApps.Metro.