Further Span optimization
- GraphNodes have an index starting at 0 set by DistanceLookup on
CalculateFully call that enables use of ints equivalent to GraphNodes as
Array indexes
- Dictonaries changed to Arrays because lookup in Dictonaries is really
slow
- Span uses ints instead of GraphNodes
- foreach loops in Span replaced by for loops because they are much
faster
- Graphs.cs clean up
The only thing that could still be optimized in Span is the
instantiation of QueueNodes that takes ~30% of Span runtime. Since I
don't see how it would work without creating new QueueNode objects,
optimization is done for now.