Affix extracting from exilemods
Done as a new project (UpdateEquipment). Will add item type extractor next.

The format is different from the one used currently, so I'll have to make changes when I'm integrating this into the program.

(Newtonsoft.Json is updated because the new project needs it and I don't know how to add a version that is not currently the newest)