Handle unknown base items on a best effort basis.
For items with unknown bases a new ItemBase is created that tries
to guess ItemGroup and ItemType from the slot, typeLine and weapon class
of the item. Fixes crashes on unknown item bases.

This is probably as far as I'll go to support Talisman and race base item types.
Talismans can at least be equipped now.