Replace Popup with MahApps like methods
Relevant for opening dialogs are the classes in Controls.Dialogs. They are mostly copied from MahApps.Metro.
Either a MetroWindow or a DialogCoordinator instance is necessary to open dialogs from a class.

The TreeGenerator's ControllerWindow and SettingsWindow have been refactored to allow for the new dialogs.
SettingsWindow is now a MahApps.Metro.ChildWindow (new NuGet package) and ControllerWindow is
a metro dialog.

These let to async/await being used far more often.

And yes, I got a bit carried away with this.