Ask user for BuildSavePath on first start
(for non-portable installations)
Should be set in installer but that is not done yet, this is necessary for update installations anyway.

As a side effect, part of the PersistentData initialization happens while the loading overlay is there and
not before as that is necessary to show custom dialogs to the user.