Prevent error border on "Ok" button
For some reason when the parameter that was validated is bound as a
CommandParamter the button gets the error border as well, and that
border don't dissapere when the border around the textbox does.

This is a fast fix, the better one would be be to create a OkCommand so
no parameters need to be provided.