Change XmlStat to allow multiple (ranged) values
This is necessary for uniques but didn't occur with base items.
Values are serialized space separated.

Added damage is no longer split into two stats (minimum and maximum)
because that is no longer necessary.

Stat.cs and everything using is it not yet adjusted, so this shouldn't
be used at this stage.