Fix a steiner reduction replacing edges with non-optimal ones
It sometimes led to incorrect reductions.

Instead of only not replacing edges with worse ones, it no longer
adds edges that are longer than the distance between the nodes.
Edges that are longer than the distance between two nodes can't
be in optimal steiner trees: They could be replaced by the shortest
path between the nodes, which would decrease the tree's size
-> the tree wasn't optimal to begin with.

Interestingly, this does increase the search space for the
known previously bugged cases, even though the test now adds less
edges.