Use IValueBuilder instead of ValueBuilder as much as possible
Remove ValueBuilder usage from all ModifierBuilding interfaces/
classes. Has the upside that == can now be used where it previously
could not (operator overloading) and that value mocks in tests can
be constructed easier (no need to wrap mocks in new ValueBuilder()).

Also, rename ModifierAggregator to ModifierResultExtensions
(the methods are now extension methods) and implement the easy
parts of it.