Fix updating from portable not installing as portable
It would install into the correct directory but use AppData for data (including updated skill trees). The installation itself would still run in portable mode because of the pre-existing portable.ini and would therefore not use the updated data.

Cause: The portable installer parameter did set values on an installer page that is skipped because of the silent installation. Directly setting IsPortable fixes that.